Add centralized configuration
