+++ bin/console
@@ -11,7 +11,6 @@ from cleo import Application
 from webdevops.docker.DockerBaseClient import DockerBaseClient
 from webdevops.docker.DockerPyClient import DockerPyClient
 from webdevops.docker.DockerCliClient import DockerCliClient
-from webdevops.doit.DoitReporter import DoitReporter
 from command.docker_build_command import DockerBuildCommand
 from command.docker_push_command import DockerPushCommand
 from command.docker_pull_command import DockerPullCommand
@@ -36,19 +35,9 @@ if __name__ == '__main__':
 
         #'dockerClient': DockerPyClient(),
         #'dockerClient': DockerCliClient(),
-        'dockerClient': DockerBaseClient(),
+        #'dockerClient': DockerBaseClient(),
 
         'blacklistFile': generatePath('BLACKLIST'),
-
-        'doitConfig': {
-            'GLOBAL': {
-                #'reporter': DoitReporter,
-               'verbosity': '0',
-           }
-       },
-
-        'dockerBuild': {},
-        'dockerPush': {},
 
         'docker': {
             'imagePrefix': 'webdevops',
